🐛 fix(Makefile): comment out unused make commands ✨ feat(Makefile): add support for recompiling the project without cleaning 🐛 fix(includes/minishell.h): change return type of rm_quotes function to t_token pointer 🐛 fix(src/debug/print_vector.c): remove unused type O_HEREDOC and SEMICOLON 🐛 fix(src/lexer/index.c): change type of i parameter in lexer_check_char function to size_t 🐛 fix(src/lexer/token.c): change err function call to set_err function call 🐛 fix(src/main.c): comment out unused variables and function calls
🔧 fix(parser/index.c): update timestamp in file header comment
🔧 fix(parser/index.c): remove unused parse_one function
🔧 fix(structs/shell.c): update timestamp in file header comment and rename signal variable to g_signal
🔧 fix(utils/miscellaneous.c): update timestamp in file header comment and fix function call to ft_strlen

+++ src/lexer/index.c
@@ -6,7 +6,7 @@
 /*   By: mdekker/jde-baai <team@codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/07/19 13:32:55 by mdekker/jde   #+#    #+#                 */
-/*   Updated: 2023/09/07 14:55:59 by mdekker/jde   ########   odam.nl         */
+/*   Updated: 2023/09/08 16:50:32 by mdekker/jde   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -93,7 +93,7 @@ static bool	check_delimiters(char *str)
 	return (true);
 }
 
-bool	lexer_check_char(char *input, int *i, t_shell *data)
+bool	lexer_check_char(char *input, size_t *i, t_shell *data)
 {
 	char	temp[2];
